move hfb_data_to_linework to faceutil module
