+++ flopy/utils/faceutil.py
@@ -281,3 +281,61 @@ def is_vertical(mg: Grid, cellid1: tuple[int, ...], cellid2: tuple[int, ...]) ->
 
         # Connection not found between cells
         raise ValueError(f"No connection found between cells {cellid1} and {cellid2}")
+
+
+def hfb_data_to_linework(recarray, modelgrid):
+    """
+    Convert HFB barrier data to line segments representing shared cell faces.
+
+    Parameters
+    ----------
+    recarray : np.recarray
+        recarray of hfb input data
+    modelgrid : Grid
+        modelgrid instance
+
+    Returns
+    -------
+    list
+        list of line segments, each as a tuple of two (x, y) coordinate tuples
+    """
+    verts = modelgrid.verts
+    nodes = []
+    if modelgrid.grid_type == "structured":
+        if "k" in recarray.dtype.names:
+            for rec in recarray:
+                node1 = modelgrid.get_node([(0, rec["irow1"], rec["icol1"])])[0]
+                node2 = modelgrid.get_node([(0, rec["irow2"], rec["icol2"])])[0]
+                nodes.append((node1, node2))
+        else:
+            for rec in recarray:
+                node1 = modelgrid.get_node([(0,) + rec["cellid1"][1:]])[0]
+                node2 = modelgrid.get_node([(0,) + rec["cellid2"][1:]])[0]
+                nodes.append((node1, node2))
+
+    elif modelgrid.grid_type == "vertex":
+        for rec in recarray:
+            nodes.append((rec["cellid1"][-1], rec["cellid2"][-1]))
+
+    else:
+        if "node1" in recarray.dtype.names:
+            nodes = list(zip(recarray["node1"], recarray["node2"]))
+        else:
+            for rec in recarray:
+                nodes.append((rec["cellid1"][0], rec["cellid2"][0]))
+
+    shared_edges = []
+    for node0, node1 in nodes:
+        edge = get_shared_face_indices(modelgrid, node0, node1)
+        if edge is None:
+            raise AssertionError(
+                f"No shared cell edges found. Cannot represent HFB "
+                f"for nodes {node0} and {node1}"
+            )
+        shared_edges.append(edge)
+
+    lines = []
+    for edge in shared_edges:
+        lines.append((tuple(verts[edge[0]]), tuple(verts[edge[1]])))
+
+    return lines
